timer panel (#93)
Creates a Timer Panel for creating the survey schedule.

The backend saves the schedule as a string in the database but the API
receives a JSON object on POST and returns one on GET. The reason a
string was used in the database was to allow for future modification of
the Timer options such as multiple times per day or even the creation of
multiple surveys.

![TimerPanel](https://github.com/user-attachments/assets/7332978f-892c-4453-90af-cba88e23a47f)

[Link to Ticket](https://incredihire-academy.atlassian.net/browse/SR-60)
[Insert related tickets here. Otherwise, delete this block]

Please include the acceptance criteria from the ticket here or select
N/A.
- [ ] N/A