SR-56 Add scheduled inquiries (frontend) (#111)
# Task
Adds ability to add inquiries to the scheduled inquiries table from the
UI.

# Author Tasks
* Added a shared SimpleModal component. This modal is good for 'Are you
sure?' type scenarios. Since it takes a React node as props, it can
serve as a modal wrapper for anything. Its up to the caller to provide
the business logic.

* I considered using the shared FormModal component for this purpose but
I opted for duplication in favor of mucking up the FormModal, which is
great for scenarios that require forms. If the team isn't in favor of
that, I can change it.

* I did not get around to building a new view for reading Scheduled
Inquiries, but the backend has been built for this and putting a view
together should be trivial.

[Link to Ticket](https://incredihire-academy.atlassian.net/browse/SR-56)
## Acceptance Criteria
* Adding a inquiry will add the inquiry to the bottom of the list.
* Inquiries can be added multiple times / Admin can repeat Inquiries. 
That is, adding an inquiry to the scheduled inquiries does not remove it
from the list of Active Inquiries. The same question can be added
multiple times.
* Save button will save the scheduled questions list to database

Please include the acceptance criteria from the ticket here or select
N/A.
- [ ] N/A