SR-16/SR-124 Playwright framework setup and create inquiries (#146)
# Task

Refactored test structure by implementing the Page Object Model and
relocating all .spec.ts files to the tests folder for better
organization. Added inquiry.spec.ts to include all updates related to
SR-16.

[Link to Ticket]

https://incredihire-academy.atlassian.net/jira/software/c/projects/SR/boards/3?selectedIssue=SR-16

[Insert related tickets here. Otherwise, delete this block]

https://incredihire-academy.atlassian.net/jira/software/c/projects/SR/boards/3?selectedIssue=SR-124

# Author Tasks

## Acceptance Criteria
As an Admin, I want to publish a set of Inquiries that will be presented
to users, to collect details on users’ state of mind, etc., so that
users and managers can align state of mind details with team and
individual activities.

ACCEPTANCE CRITERIA

Admin portal should have a button that allows the admin to add an
inquiry

“Add Inquiry” Button should trigger a pop-up window that displays an
input box labeled "Enter the text of your inquiry."

Style the input box and question list to ensure a responsive and
user-friendly interface.

error messages under the inputs for validation

toast notification for network or backend errors to cover the necessary
feedback.

Pop-up window should have a SAVE and CANCEL button

The SAVE button closes the window and displays a success message

The CANCEL button or the X closes the window, and no inquiry is added to
the database.

The system should store the inquiries in the database to prevent the
entry of duplicate questions.